* origin/master:
  fix: merge hazard from flagWithMetadata PR
  cleanup: reorder setParam function, uint256 -> uint in Policies
  feat: make slashing fraction configurable (#624)
  rename flagWithMetadata -> flag, and remove the previous implementation (#630)

+++ packages/network-contracts/contracts/OperatorTokenomics/SponsorshipPolicies/IJoinPolicy.sol
@@ -3,6 +3,6 @@
 pragma solidity ^0.8.13;
 
 interface IJoinPolicy {
-    function setParam(uint256 param) external;
+    function setParam(uint param) external;
     function onJoin(address operator, uint amount) external;
 }
